backend/backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
openai-whisper
torch
numpy
